@@ -12,48 +12,49 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.2",
-    "@angular-devkit/core": "0.0.28",
-    "@angular/animations": "^5.2.3",
-    "@angular/cdk": "^5.1.1",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.2.1",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "^5.1.1",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.9",
+    "@agm/core": "^1.0.0-beta.3",
+    "@angular/animations": "^6.1.0",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^2.2.1",
     "angular-font-awesome": "^3.1.2",
-    "angular-in-memory-web-api": "^0.5.3",
-    "bootstrap": "^4.0.0",
-    "core-js": "^2.4.1",
+    "angular-in-memory-web-api": "^0.6.1",
+    "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "rxjs-compat": "^6.2.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.6.5",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular-devkit/build-angular": "^0.7.1",
+    "@angular-devkit/core": "0.7.1",
+    "@angular/cli": "^6.1.1",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
     "@types/googlemaps": "^3.30.11",
-    "@types/jasmine": "~2.5.53",
+    "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@types/node": "~10.5.4",
+    "codelyzer": "^4.4.2",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.5",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^1.2.0",
+    "protractor": "~5.4.0",
+    "rxjs": "^6.2.2",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "^2.9.2",
+    "zone.js": "^0.8.26"
   }
 }
