@@ -1,6 +1,7 @@
+/// <reference types="googlemaps" />
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import {} from '@types/googlemaps';
+// import {} from 'googlemaps';
 
 @Component({
   selector: 'app-login',
